chore: remove define keymap in service
