+++ packages/blocks/src/__internal__/utils/common-operations.ts
@@ -2,7 +2,7 @@ import { assertExists, matchFlavours } from '@blocksuite/global/utils';
 import type { Page } from '@blocksuite/store';
 import type { BaseBlockModel } from '@blocksuite/store';
 import type { Workspace } from '@blocksuite/store';
-import type { VEditor, VRange } from '@blocksuite/virgo';
+import type { VRange } from '@blocksuite/virgo';
 
 import type { ListType } from '../../list-block/index.js';
 import { asyncGetRichTextByModel, getVirgoByModel } from './query.js';
@@ -28,16 +28,6 @@ export function asyncFocusRichText(
   assertExists(model);
   if (matchFlavours(model, ['affine:divider'])) return;
   return asyncSetVRange(model, vRange);
-}
-
-export function isCollapsedAtBlockStart(vEditor: VEditor) {
-  const vRange = vEditor.getVRange();
-  return vRange?.index === 0 && vRange?.length === 0;
-}
-
-export function isCollapsedAtBlockEnd(vEditor: VEditor) {
-  const vRange = vEditor.getVRange();
-  return vRange?.index === vEditor.yText.length && vRange?.length === 0;
 }
 
 export function isInSamePath(
