+++ packages/editor/src/components/editor-container.ts
@@ -144,18 +144,6 @@ export class EditorContainer
         this.requestUpdate('page');
       })
     );
-    this._disposables.add(
-      this.page.slots.blockUpdated.on(async ({ type, id }) => {
-        const block = this.page.getBlockById(id);
-
-        if (!block) return;
-
-        if (type === 'update') {
-          const service = await getServiceOrRegister(block.flavour);
-          service.updateEffect(block);
-        }
-      })
-    );
 
     this._disposables.addFromEvent(
       this,
